world loads in circles

+++ VoxelCraft/vecXZ.h
@@ -3,7 +3,7 @@
 #include <functional>
 
 struct VecXZ {
-	std::int16_t x, z;
+	std::int64_t x, z;
 };
 
 bool operator==(const VecXZ& lhs, const VecXZ& rhs);
@@ -13,7 +13,7 @@ namespace std {
 	template<>
 	struct hash<VecXZ> {
 		size_t operator()(const VecXZ& vec) const {
-			std::hash<std::int16_t> hasher;
+			std::hash<std::int64_t> hasher;
 			auto hashx = hasher(vec.x);
 			auto hashz = hasher(vec.z);
